refactor: split native C collector into per-metric backends
Replace monolithic collector.c with a structured layout mirroring the
Python metrics hierarchy:

  native_c/
    monitor.h          — shared TickContext + CCollector interface
    monitor.c          — main(), registry, emit_ready/emit_tick
    metrics/cpu/       — /proc/<pid>/stat + /proc/stat
    metrics/memory/    — /proc/<pid>/statm + /proc/meminfo
    metrics/io/        — /proc/<pid>/io + /proc/diskstats
    metrics/gpu/       — NVML via dynamic loading

Active collectors are selected via --collectors arg driven by
config/c/collectors.yaml, loaded by CSubprocessPerformanceMonitor.
Makefile updated to compile all five translation units via make.
