+++ .gitignore
@@ -173,8 +173,9 @@ poetry.toml
 # LSP config files
 pyrightconfig.json
 
-# Compiled C monitoring collector
+# Compiled C monitoring collector and build artifacts
 jumper_collector
+*.o
 
 # Benchmark results
 jumper_extension/monitor/benchmark/results/
