+++ jumper_extension/config/c/collectors.yaml
@@ -1,0 +1,5 @@
+collectors:
+  - cpu
+  - memory
+  - io
+  - gpu
