+++ jumper_extension/monitor/backends/native_c/metrics/io/io.h
@@ -1,0 +1,8 @@
+#ifndef JUMPER_METRICS_IO_H
+#define JUMPER_METRICS_IO_H
+
+#include "monitor.h"
+
+extern const CCollector io_collector;
+
+#endif /* JUMPER_METRICS_IO_H */
