+++ jumper_extension/monitor/backends/native_c/metrics/memory/memory.h
@@ -1,0 +1,8 @@
+#ifndef JUMPER_METRICS_MEMORY_H
+#define JUMPER_METRICS_MEMORY_H
+
+#include "monitor.h"
+
+extern const CCollector memory_collector;
+
+#endif /* JUMPER_METRICS_MEMORY_H */
